Refactor main navigation items
Whilst the previous commit (#a87fbaa) positioned the main navigation
items, upon revisting it the implementation didn't seem to be the best
approach as it utilised containing divs rather than applying styling to
individual list items.

This commit updates how the main navigation items are positioned, by
adding styling to the list items and using different classes depending
on the context of the menu item. This results in code which is easier to
understand, and allows for more granular styling changes in future if
required.